Move url out of shared dir
